pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "phycam-eval"
version = "0.1.0"
description = "Physics-Inspired Camera Degradation Benchmarks for Object Detection Robustness"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [
    { name = "Parthiv Nair", email = "parthiv.nair@gmail.com" }
]
keywords = [
    "camera",
    "perception",
    "robustness",
    "fourier-optics",
    "degradation",
    "benchmark",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering :: Image Processing",
    "Programming Language :: Python :: 3",
]

dependencies = [
    "torch>=2.1",
    "numpy>=1.24",
    "scipy>=1.11",
    "matplotlib>=3.7",
    "pillow>=10.0",
    "tqdm>=4.65",
    "pycocotools>=2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "pytest-cov>=4.1",
    "black>=23.0",
    "ruff>=0.1",
]
viz = [
    "seaborn>=0.12",
    "pandas>=2.0",
]
experiments = [
    "ultralytics>=8.3",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["phycam_eval*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"

[tool.ruff]
line-length = 100
select = ["E", "F", "W", "I"]

[tool.black]
line-length = 100
target-version = ["py310"]
